requirements.txt
discord.py 
python-dotenv 
supabase 
yfinance 
pycoingecko 
tabulate